Fix integer overflow in DIV and DIVU

+++ Interpreter Ops.c
@@ -800,8 +800,8 @@ void _fastcall r4300i_SPECIAL_MULTU (void) {
 
 void _fastcall r4300i_SPECIAL_DIV (void) {
 	if ( GPR[Opcode.BRANCH.rt].UDW != 0 ) {
-		LO.DW = GPR[Opcode.BRANCH.rs].W[0] / GPR[Opcode.BRANCH.rt].W[0];
-		HI.DW = GPR[Opcode.BRANCH.rs].W[0] % GPR[Opcode.BRANCH.rt].W[0];
+		LO.DW = (long long)GPR[Opcode.BRANCH.rs].W[0] / (long long)GPR[Opcode.BRANCH.rt].W[0];
+		HI.DW = (long long)GPR[Opcode.BRANCH.rs].W[0] % (long long)GPR[Opcode.BRANCH.rt].W[0];
 	} else {
 		if (ShowDebugMessages)
 			DisplayError("DIV by 0 ???");
@@ -810,8 +810,8 @@ void _fastcall r4300i_SPECIAL_DIV (void) {
 
 void _fastcall r4300i_SPECIAL_DIVU (void) {
 	if ( GPR[Opcode.BRANCH.rt].UDW != 0 ) {
-		LO.DW = (int)(GPR[Opcode.BRANCH.rs].UW[0] / GPR[Opcode.BRANCH.rt].UW[0]);
-		HI.DW = (int)(GPR[Opcode.BRANCH.rs].UW[0] % GPR[Opcode.BRANCH.rt].UW[0]);
+		LO.DW = (long long)GPR[Opcode.BRANCH.rs].UW[0] / (long long)GPR[Opcode.BRANCH.rt].UW[0];
+		HI.DW = (long long)GPR[Opcode.BRANCH.rs].UW[0] % (long long)GPR[Opcode.BRANCH.rt].UW[0];
 	} else {
 		if (ShowDebugMessages)
 			DisplayError("DIVU by 0 ???");
